# faster implementation of method `sage.graphs.generators.families.BalancedTree`